Add LNY 2025 stand
The assignment banner for the previous year stuff will come at a later date or not at all. It's not where it's supposed to be (`UI\Art\Job_Terminal`) and I don't feel like spending hours on looking through all the game files to find a rogue assignment banner for a one-time job.